awssqs: move reconciler
