+++ contrib/awssqs/pkg/reconciler/resources/receive_adapter.go
@@ -19,7 +19,7 @@ package resources
 import (
 	"fmt"
 
-	"github.com/knative/eventing-sources/pkg/apis/sources/v1alpha1"
+	"github.com/knative/eventing-sources/contrib/awssqs/pkg/apis/sources/v1alpha1"
 	v1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
